requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-colorfield==0.8.0
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.4
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
